Evarutil.finalize: combine minimize, to_constr and restrict.
If you look at the context of the changes you may notice some
Pretyping.check_evars calls. This is in some way redundant with
to_constr's abort_on_undefined_evars, and is sometimes called after
a non-evar-permissive to_constr (indicating a probable bug introduced
by the abort flag).

I didn't fix any of those, leaving the order as-is. It should be done
in a separate PR, but we may want to consider how to deal with the
redundancy in a more principled manner. Maybe to_constr should be
raising a sensible error, not an anomaly? There is also
flush_and_check_evars which is used only once.